+++ packages/gatsby-source-wordpress/src/normalize.js
@@ -60,14 +60,18 @@ exports.normalizeACF = normalizeACF
 exports.combineACF = function(entities) {
   let acfOptionData = {}
   // Map each ACF Options object keys/data to single object
-  _.forEach(entities.filter(e => e.__type === `wordpress__acf_options`), e => {
-    if (e[`acf`]) {
-      acfOptionData[e.__acfOptionPageId || `options`] = {}
-      Object.keys(e[`acf`]).map(
-        k => (acfOptionData[e.__acfOptionPageId || `options`][k] = e[`acf`][k])
-      )
+  _.forEach(
+    entities.filter(e => e.__type === `wordpress__acf_options`),
+    e => {
+      if (e[`acf`]) {
+        acfOptionData[e.__acfOptionPageId || `options`] = {}
+        Object.keys(e[`acf`]).map(
+          k =>
+            (acfOptionData[e.__acfOptionPageId || `options`][k] = e[`acf`][k])
+        )
+      }
     }
-  })
+  )
 
   // Remove previous ACF Options objects (if any)
   _.pullAll(
@@ -312,12 +316,19 @@ exports.mapPolylangTranslations = entities =>
   entities.map(entity => {
     if (entity.polylang_translations) {
       entity.polylang_translations___NODE = entity.polylang_translations.map(
-        translation =>
-          entities.find(
+        translation => {
+          const post = entities.find(
             t =>
               t.wordpress_id === translation.wordpress_id &&
               entity.__type === t.__type
-          ).id
+          )
+
+          if (!post) {
+            return null
+          }
+
+          return post.id
+        }
       )
 
       delete entity.polylang_translations