+++ packages/gatsby/src/bootstrap/requires-writer.js
@@ -2,11 +2,42 @@ const _ = require(`lodash`)
 const path = require(`path`)
 const fs = require(`fs-extra`)
 const crypto = require(`crypto`)
-const slash = require(`slash`)
+const { slash } = require(`gatsby-core-utils`)
 const { store, emitter } = require(`../redux/`)
 const reporter = require(`gatsby-cli/lib/reporter`)
 const { match } = require(`@reach/router/lib/utils`)
 import { joinPath } from "gatsby-core-utils"
+
+// path ranking algorithm copied (with small adjustments) from `@reach/router` (internal util, not exported from the package)
+// https://github.com/reach/router/blob/28a79e7fc3a3487cb3304210dc3501efb8a50eba/src/lib/utils.js#L216-L254
+const paramRe = /^:(.+)/
+
+const SEGMENT_POINTS = 4
+const STATIC_POINTS = 3
+const DYNAMIC_POINTS = 2
+const SPLAT_PENALTY = 1
+const ROOT_POINTS = 1
+
+const isRootSegment = segment => segment === ``
+const isDynamic = segment => paramRe.test(segment)
+const isSplat = segment => segment === `*`
+
+const rankRoute = path =>
+  segmentize(path).reduce((score, segment) => {
+    score += SEGMENT_POINTS
+    if (isRootSegment(segment)) score += ROOT_POINTS
+    else if (isDynamic(segment)) score += DYNAMIC_POINTS
+    else if (isSplat(segment)) score -= SEGMENT_POINTS + SPLAT_PENALTY
+    else score += STATIC_POINTS
+    return score
+  }, 0)
+
+const segmentize = uri =>
+  uri
+    // strip starting/ending slashes
+    .replace(/(^\/+|\/+$)/g, ``)
+    .split(`/`)
+// end of copied `@reach/router` internals
 
 let lastHash = null
 
@@ -29,19 +60,10 @@ const getComponents = pages =>
  */
 const getMatchPaths = pages => {
   const createMatchPathEntry = (page, index) => {
-    let score = page.matchPath.replace(/[/][*]?$/, ``).split(`/`).length
-    let wildcard = 0
-
-    if (!page.matchPath.includes(`*`)) {
-      wildcard = 1
-      score += 1
-    }
-
     return {
       ...page,
       index,
-      score,
-      wildcard,
+      score: rankRoute(page.matchPath),
     }
   }
 
@@ -53,11 +75,12 @@ const getMatchPaths = pages => {
   })
 
   // Pages can live in matchPaths, to keep them working without doing another network request
-  // we save them in matchPath. Our sorting will put them above dynamic routes
-  // as we add a static bonus point to static routes
+  // we save them in matchPath. We use `@reach/router` path ranking to score paths/matchPaths
+  // and sort them so more specific paths are before less specific paths.
   // More info in https://github.com/gatsbyjs/gatsby/issues/16097
   // small speedup: don't bother traversing when no matchPaths found.
   if (matchPathPages.length) {
+    const newMatches = []
     pages.forEach((page, index) => {
       const isInsideMatchPath = !!matchPathPages.find(
         pageWithMatchPath =>
@@ -65,7 +88,7 @@ const getMatchPaths = pages => {
       )
 
       if (isInsideMatchPath) {
-        matchPathPages.push(
+        newMatches.push(
           createMatchPathEntry(
             {
               ...page,
@@ -76,24 +99,21 @@ const getMatchPaths = pages => {
         )
       }
     })
+    // Add afterwards because the new matches are not relevant for the existing search
+    matchPathPages.push(...newMatches)
   }
 
   return matchPathPages
     .sort((a, b) => {
-      // Paths with wildcards should appear after those without.
-      const wildcardOrder = b.wildcard - a.wildcard
-      if (wildcardOrder !== 0) {
-        return wildcardOrder
-      }
-
       // The higher the score, the higher the specificity of our matchPath
       const order = b.score - a.score
       if (order !== 0) {
         return order
       }
 
-      // if specificity is the same we use the array index
-      return b.index - a.index
+      // if specificity is the same we do lexigraphic comparison of path to ensure
+      // deterministic order regardless of order pages where created
+      return a.matchPath.localeCompare(b.matchPath)
     })
     .map(({ path, matchPath }) => {
       return { path, matchPath }
@@ -153,7 +173,7 @@ const preferDefault = m => m && m.default || m
       )
 
       return `  "${c.componentChunkName}": () => import("${slash(
-        relativeComponentPath
+        `./${relativeComponentPath}`
       )}" /* webpackChunkName: "${c.componentChunkName}" */)`
     })
     .join(`,\n`)}