+++ www/src/components/layout.js
@@ -3,11 +3,11 @@ import { jsx } from "theme-ui"
 import React from "react"
 import { navigate, PageRenderer } from "gatsby"
 import mousetrap from "mousetrap"
-import Modal from "react-modal"
 import MdClose from "react-icons/lib/md/close"
 
 import { Global } from "@emotion/core"
 
+import { getItemList } from "../utils/sidebar/item-list"
 import { globalStyles } from "../utils/styles/global"
 import {
   colors,
@@ -24,8 +24,12 @@ import PageWithSidebar from "../components/page-with-sidebar"
 import SiteMetadata from "../components/site-metadata"
 import SkipNavLink from "../components/skip-nav-link"
 import "../assets/fonts/futura"
+import LazyModal from "./lazy-modal"
+import { defaultLang } from "../utils/i18n"
 
 let windowWidth
+
+export const LocaleContext = React.createContext(defaultLang)
 
 class DefaultLayout extends React.Component {
   constructor() {
@@ -38,8 +42,6 @@ class DefaultLayout extends React.Component {
   }
 
   componentDidMount() {
-    Modal.setAppElement(`#___gatsby`)
-
     if (this.props.isModal && window.innerWidth > 750) {
       mousetrap.bind(`left`, this.props.modalPrevious)
       mousetrap.bind(`right`, this.props.modalNext)
@@ -60,9 +62,9 @@ class DefaultLayout extends React.Component {
   }
 
   render() {
+    const itemList = getItemList(this.props.location.pathname)
     // SEE: template-docs-markdown for why this.props.isSidebarDisabled is here
-    const isSidebarDisabled =
-      this.props.isSidebarDisabled || !this.props.itemList
+    const isSidebarDisabled = this.props.isSidebarDisabled || !itemList
     let isModal = false
     if (!windowWidth && typeof window !== `undefined`) {
       windowWidth = window.innerWidth
@@ -79,7 +81,7 @@ class DefaultLayout extends React.Component {
           <PageRenderer
             location={{ pathname: this.props.modalBackgroundPath }}
           />
-          <Modal
+          <LazyModal
             isOpen={true}
             style={{
               content: {
@@ -161,15 +163,18 @@ class DefaultLayout extends React.Component {
               {this.props.modalPreviousLink}
               {this.props.modalNextLink}
             </div>
-          </Modal>
+          </LazyModal>
         </>
       )
     }
 
     return (
-      <>
+      <LocaleContext.Provider value={this.props.locale || defaultLang}>
         <Global styles={globalStyles} />
-        <SiteMetadata pathname={this.props.location.pathname} />
+        <SiteMetadata
+          pathname={this.props.location.pathname}
+          locale={this.props.locale}
+        />
         <SkipNavLink />
         <Banner />
         <Navigation pathname={this.props.location.pathname} />
@@ -189,14 +194,14 @@ class DefaultLayout extends React.Component {
         >
           <PageWithSidebar
             disable={isSidebarDisabled}
-            itemList={this.props.itemList}
+            itemList={itemList}
             location={this.props.location}
             enableScrollSync={this.props.enableScrollSync}
             renderContent={() => this.props.children}
           />
         </div>
         <MobileNavigation />
-      </>
+      </LocaleContext.Provider>
     )
   }
 }