+++ www/src/pages/contributing/events.js
@@ -3,7 +3,6 @@ import { Helmet } from "react-helmet"
 import { graphql } from "gatsby"
 
 import Layout from "../../components/layout"
-import { itemListContributing } from "../../utils/sidebar/item-list"
 import Container from "../../components/container"
 import EmailCaptureForm from "../../components/email-capture-form"
 import DocSearchContent from "../../components/docsearch-content"
@@ -11,7 +10,7 @@ import FooterLinks from "../../components/shared/footer-links"
 import Events from "../../components/events/events"
 
 const EventsRoute = props => (
-  <Layout location={props.location} itemList={itemListContributing}>
+  <Layout location={props.location}>
     <DocSearchContent>
       <Container>
         <Helmet>
@@ -26,9 +25,7 @@ const EventsRoute = props => (
         </h1>
         <p>
           Interested in connecting with the Gatsby community in person? Take a
-          look at the events below to see community organized Gatsby events,
-          places the Gatsby team members are speaking, and conferences that
-          Gatsby is sponsoring. We hope to see you at these events soon!
+          look at the list below to see community-organized Gatsby events.
         </p>
         <p>
           Want to see your event featured here?